Moved most items from doubles to floats, they didn't need that precison.
git-svn-id: https://www.cs.binghamton.edu/svn/simple@930 bb5c8b44-a00f-0410-a998-867a3eb9c5a3

+++ simplegui/sg.cpp
@@ -34,7 +34,7 @@ using namespace std;
 
 SimpleGUI *current_sg = NULL;
 
-SimpleGUI::SimpleGUI(int aspmeth, const double asp) {
+SimpleGUI::SimpleGUI(int aspmeth, const float asp) {
   if(current_sg) {
     fprintf(stderr, "ERROR: Attempted to create multiple SimpleGUI instances.\n");
     exit(1);
