+++ simplegui/sg_table.h
@@ -57,7 +57,7 @@ public:
   virtual void RemoveRows(const vector<int> &yels);
   virtual void Substitute(SG_Widget *oldwid, SG_Widget *newwid);
 
-  virtual void SetAspectRatio(double asp);
+  virtual void SetAspectRatio(float asp);
 
   void LinkResize(SG_Ranger2D *ranger);
   void LinkXResize(SG_Ranger2D *ranger);
